Fix WebXR display issue for angled screens when using OpenVR
Sending the full head-from-eye transformation matrix instead of just
the translation component and then multiplying the view matrix by it
fixesthe issue of having double vision when using a headset such as the
pimax 5k with significantly angled screens through OpenVR.

Also updated naming of transforms in XRView to be more readable.

Platforms that now send the actual head-from-eye transform:
- OpenVR
- GVR

Platforms that still only send the offset, but have at least been
updated to send it via the transformation matrix. These can be updated
in separate CLs.
- WMR
- Oculus (getting full transform would require updating libovr version)
- OpenXR

Because WebVR blink code uses the same mojo structs, it required a
small update to extract the translation component from the
transformation matrix that it receives. I verified that the WebVR
behavior is still the same (doesn't work with angled screens before or
after this change, but non-angled screens work fine).

Bug: 928433
Change-Id: Icf1004bb6ced033c5e4d68bf80686aeb77c393a6